refactor: split up mission selectors
